requirements.txt
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
decorator==5.2.1
idna==3.11
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
msgpack==1.1.2
numba==0.63.1
numpy==2.2.6
packaging==26.0
pedalboard==0.9.22
platformdirs==4.6.0
pooch==1.9.0
pycparser==3.0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
soundfile==0.13.1
soxr==1.0.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
